test: consolidate goal store suites
